phase 1/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0

# Async task queue
celery==5.4.0
redis==5.0.4
arq==0.26.0

# Document parsing
pypdf==4.2.0
python-docx==1.1.2
python-multipart==0.0.9
httpx==0.27.0

# Chunking
langchain-text-splitters==0.2.1
tiktoken==0.7.0

# Embeddings
openai==1.35.0

# Vector store (choose one)
qdrant-client==1.9.1
# weaviate-client==4.6.1

# Observability
mlflow==2.14.1
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
prometheus-fastapi-instrumentator==7.0.0

# Utils
structlog==24.2.0
tenacity==8.3.0
python-dotenv==1.0.1
